These are some basic c++ programs I have made.
to upper and camel create upper case text in terminal windows. 
test is a simple hello world and check input. 
fizzbuzz is my take on playing fizzbuzz with zero players. You tell the computer how high you would like it to go and it goes. 
any file ending in X is experimental.